[code] When using the serial interface, allow entering the machine interface after the terminal console has been started.
A read functinoid has been added that detects the netx 90 terminal console (HBoot 3).
The Connect() method has been extended to try to determine the chip type using this functinoid.
When successful, the update method of this functinoid is called, which will send a boot image using the usip command.
This boot image may, e.g., force the console mode. After executing the boot image, if the netx ends up in console mode again, it will be listening for a knock sequence, allowing us to enter the machine interface.
Alternatively, the boot image may contain a RAM version of the machine interface.

In order to pass this boot image to the romloader, an option structure has been introduced to romloader_uart.
It contains the attribute "netx90_m2m_image". The value is a boot image in binary format (not uuencoded) that can be sent via the usip command.