[code] When using the serial interface, allow entering the machine interface after the terminal console has been started.
When a netx 90 in non-secure boot mode is found and the terminal protocol of the console is active, send a reset command and start the machine interface.

A read functinoid has been added that supports the terminal protocol of the netx 90 console mode.
- detect_console_mode() uses the command "s 0 4" to check if the console is in open or secure mode.
- read_data32() uses the d(ump) command to detect the chip type
- update_device uses the reset command to bring the netx back to the start of the UART protocol,
where it will enter the machine interface on the appropriate knock sequence.

The Connect() method has been updated to try the netx 90 terminal protocol (hboot3) when the
previously tried terminal protocols were unsuccessful.
- detect if the console is in open or secure mode (i.e. if secure boot is off or on).
- only continue in open mode, otherwise exit with an error.
- detect the chip type
- if the chip type is netx 90, send a reset command.

Limitations:
- Only the non-secure boot mode is supported.
- The console mode must be reached, either using the console mode jumper, or by making sure that no boot image is present.

Note: The netx 4000 understands a similar protocol, but does not support the reset command.
The chip type will be detected, but the chip cannot be reset, so it is not possible to switch
from the terminal protocol to the machine interface in this way.