[code] Fix detection of netx 90 rev.1
Instead of checking the netx version register (0xff401298) for the value 0x0900010d, check location 0x5110 in the ROM for the value 0xe001200c.
The netx version register is not yet set when we use JTAG and halt at 0x170.
This is needed to discriminate between netx 90 rev.1 and rev.2.