[hotfix] Fix crash in get_info() on Ubuntu (NXTFLASHER-534)
get_info() used an incorrect packet size for the command_info packet. The packet size passed to send_packet() was too large, leading to stack corruption because send_packet() puts a checksum at the end of the packet.
For some reason, this does not cause a crash on Windows. The packet is accepted by the monitor, because it does not check if the packet is larger than expected.