[code] Update chip type detection and starting machine interface from terminal console (NXTFLASHER-568)
Change the process of connecting to a netX that is in terminal console mode:
detect secure mode
if open, detect the chip type using dump commands
if secure, check if the dump command works.
If it does, detect the chip type using dump commands
If it does not, send the VERS command.
If the output is 05/0d, assume that the chip type is netx 90 Rev.2.

When a boot image is sent to the terminal mode console in order to start the machine interface:
return an error code if execution of the image fails for some reason.
Accept a MI sync packet with MI version 3.x