[code] Add the function get_console_mode (NXTFLASHER-577)
The console mode (open, secure or unknown) is stored during a connection attempt and can be retrieved from Lua using get_console_mode().
When a terminal mode console is found, the console mode is determined by trying to execute the command s 0 4.
When MI3.1 is active, the info request is sent and the console mode is taken from the response.
When MI3.0 is active, the console mode cannot be determined and is returned as unknown.
This works only for romloader_uart.