[code] Update romloader test for lua 5.4
test_romloader.lua: load the "bit" package in a way that is compatible with Lua 5.4. This script requires the "bit" package and will work with both Lua 5.1 and 5.4.
bit.lua: This is the Lua 5.4 compatible replacement for the luajit bit package.
test_romloader_lua54.lua: This version does not use bit operations, instead it uses string.pack.